Fix finding node address with duplicate type - Closes #3458
